services/message-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
aiokafka==0.12.0
structlog==25.1.0
python-dotenv==1.1.0
alembic==1.15.2
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
ruff==0.11.2
